requirements.txt
tensorflow
scikit-learn
matplotlib
Pillow
numpy
pandas